PUT on categories implemented

+++ YG_server/categories/routes.py
@@ -1,6 +1,8 @@
 from flask import Blueprint, jsonify, request, abort
 from datetime import datetime as dt
 
+from flask.helpers import url_for
+
 from .models import db, Category
 from YG_server.users.models import User
 
@@ -13,8 +15,10 @@ def all():
     name = created_category['name']
     user_id = created_category['user_id']
 
-    if name is None or user_id is None:
-      abort(400, description="Name not passed from client")
+    if user_id is None:
+      abort(400, description="user_id not passed from client")
+    if name is None: 
+      abort(400, description="name not passed from client")
     
     user = User.query.get(user_id)
     if user is None:
@@ -32,15 +36,22 @@ def all():
     db.session.add(new_category)
     db.session.commit()
 
-
-    return jsonify(f'category created: {new_category.name}'), 201
+    return jsonify({'name': f'{new_category.name}'}), 201
 
 
   categories = Category.query.all()
   if categories is None:
     abort(404, description="No categories available")
 
-  return jsonify(str(categories)), 200
+  res = []
+  for category in categories:
+    res.append({
+      'title': f'{category.name}',
+      'uri': f'http://localhost:5000/api/v1.0/categories/{category.id}',
+      # 'uri': url_for('get_cateogry', category_id = f'{category.id}')
+    })
+
+  return jsonify(res), 200
 
 @categories_bp.route('/<int:category_id>', methods=['GET', 'DELETE', 'PUT'])
 def get_category(category_id):
@@ -50,10 +61,25 @@ def get_category(category_id):
   if found_category is None:
     abort(404, description="Category does not exist")
 
+  if request.method == 'PUT':
+    new_name = request.get_json()['name']
+    found_category.name = new_name
+    db.session.commit()
+    return jsonify({
+      'name': f'{found_category.name}', 
+      'uri': f'http://localhost:5000/api/v1.0/categories/{found_category.id}',
+      # 'created': f'{found_category.created}'
+    })
+
   if request.method == 'DELETE':
     db.session.delete(found_category)
     db.session.commit()
     return jsonify({'message': f'category \'{found_category.name}\' has been deleted'})
 
-  return jsonify({'message': f'category: {found_category}'})
+  # instead of returning category id return URI
+  return jsonify({
+    'name': f'{found_category.name}', 
+    'uri': f'http://localhost:5000/api/v1.0/categories/{found_category.id}',
+    # 'created': f'{found_category.created}'
+  })
